requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
numpy==1.24.2
pandas==1.5.3
pymongo==3.12.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.2
six==1.16.0
sqlparse==0.2.4
tzdata==2023.2
